Move the CompletableFuture config to the view model
